spi: spidev: add novena to compatible list
The Novena FPGA is connected via SPI, and being an FPGA, it can
be reconfigured to be anything.  Therefore, it is easiest to
treat it as a generic SPI device, accessible from userspace.

Add kosagi,novena-fpga-spidev to the "compatible" list in order
to allow Novena to boot with a spidev device without generating
a warning on boot.

Signed-off-by: Sean Cross <xobs@kosagi.com>